+++ sdk/containerservicefleet/azure-resourcemanager-containerservicefleet/src/main/java/com/azure/resourcemanager/containerservicefleet/fluent/GatesClient.java
@@ -134,6 +134,9 @@ public interface GatesClient {
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param fleetName The name of the Fleet resource.
+     * @param filter Filter the result list using the given expression.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -141,5 +144,6 @@ public interface GatesClient {
      * @return the response of a Gate list operation as paginated response with {@link PagedIterable}.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    PagedIterable<GateInner> listByFleet(String resourceGroupName, String fleetName, Context context);
+    PagedIterable<GateInner> listByFleet(String resourceGroupName, String fleetName, String filter, Integer top,
+        String skipToken, Context context);
 }