+++ sdk/containerservicefleet/azure-resourcemanager-containerservicefleet/src/main/java/com/azure/resourcemanager/containerservicefleet/implementation/AutoUpgradeProfilesClientImpl.java
@@ -140,6 +140,7 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
         Mono<Response<AutoUpgradeProfileListResult>> listByFleet(@HostParam("endpoint") String endpoint,
             @QueryParam("api-version") String apiVersion, @PathParam("subscriptionId") String subscriptionId,
             @PathParam("resourceGroupName") String resourceGroupName, @PathParam("fleetName") String fleetName,
+            @QueryParam("$top") Integer top, @QueryParam("$skipToken") String skipToken,
             @HeaderParam("Accept") String accept, Context context);
 
         @Headers({ "Content-Type: application/json" })
@@ -149,6 +150,7 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
         Response<AutoUpgradeProfileListResult> listByFleetSync(@HostParam("endpoint") String endpoint,
             @QueryParam("api-version") String apiVersion, @PathParam("subscriptionId") String subscriptionId,
             @PathParam("resourceGroupName") String resourceGroupName, @PathParam("fleetName") String fleetName,
+            @QueryParam("$top") Integer top, @QueryParam("$skipToken") String skipToken,
             @HeaderParam("Accept") String accept, Context context);
 
         @Headers({ "Content-Type: application/json" })
@@ -765,6 +767,8 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param fleetName The name of the Fleet resource.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
@@ -773,14 +777,33 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     private Mono<PagedResponse<AutoUpgradeProfileInner>> listByFleetSinglePageAsync(String resourceGroupName,
-        String fleetName) {
+        String fleetName, Integer top, String skipToken) {
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.listByFleet(this.client.getEndpoint(), this.client.getApiVersion(),
-                this.client.getSubscriptionId(), resourceGroupName, fleetName, accept, context))
+                this.client.getSubscriptionId(), resourceGroupName, fleetName, top, skipToken, accept, context))
             .<PagedResponse<AutoUpgradeProfileInner>>map(res -> new PagedResponseBase<>(res.getRequest(),
                 res.getStatusCode(), res.getHeaders(), res.getValue().value(), res.getValue().nextLink(), null))
             .contextWrite(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext()).readOnly()));
+    }
+
+    /**
+     * List AutoUpgradeProfile resources by Fleet.
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param fleetName The name of the Fleet resource.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the response of a AutoUpgradeProfile list operation as paginated response with {@link PagedFlux}.
+     */
+    @ServiceMethod(returns = ReturnType.COLLECTION)
+    private PagedFlux<AutoUpgradeProfileInner> listByFleetAsync(String resourceGroupName, String fleetName, Integer top,
+        String skipToken) {
+        return new PagedFlux<>(() -> listByFleetSinglePageAsync(resourceGroupName, fleetName, top, skipToken),
+            nextLink -> listByFleetNextSinglePageAsync(nextLink));
     }
 
     /**
@@ -795,7 +818,9 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
     private PagedFlux<AutoUpgradeProfileInner> listByFleetAsync(String resourceGroupName, String fleetName) {
-        return new PagedFlux<>(() -> listByFleetSinglePageAsync(resourceGroupName, fleetName),
+        final Integer top = null;
+        final String skipToken = null;
+        return new PagedFlux<>(() -> listByFleetSinglePageAsync(resourceGroupName, fleetName, top, skipToken),
             nextLink -> listByFleetNextSinglePageAsync(nextLink));
     }
 
@@ -804,17 +829,20 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param fleetName The name of the Fleet resource.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the response of a AutoUpgradeProfile list operation along with {@link PagedResponse}.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    private PagedResponse<AutoUpgradeProfileInner> listByFleetSinglePage(String resourceGroupName, String fleetName) {
+    private PagedResponse<AutoUpgradeProfileInner> listByFleetSinglePage(String resourceGroupName, String fleetName,
+        Integer top, String skipToken) {
         final String accept = "application/json";
         Response<AutoUpgradeProfileListResult> res
             = service.listByFleetSync(this.client.getEndpoint(), this.client.getApiVersion(),
-                this.client.getSubscriptionId(), resourceGroupName, fleetName, accept, Context.NONE);
+                this.client.getSubscriptionId(), resourceGroupName, fleetName, top, skipToken, accept, Context.NONE);
         return new PagedResponseBase<>(res.getRequest(), res.getStatusCode(), res.getHeaders(), res.getValue().value(),
             res.getValue().nextLink(), null);
     }
@@ -824,6 +852,8 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param fleetName The name of the Fleet resource.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws ManagementException thrown if the request is rejected by server.
@@ -832,11 +862,11 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     private PagedResponse<AutoUpgradeProfileInner> listByFleetSinglePage(String resourceGroupName, String fleetName,
-        Context context) {
+        Integer top, String skipToken, Context context) {
         final String accept = "application/json";
         Response<AutoUpgradeProfileListResult> res
             = service.listByFleetSync(this.client.getEndpoint(), this.client.getApiVersion(),
-                this.client.getSubscriptionId(), resourceGroupName, fleetName, accept, context);
+                this.client.getSubscriptionId(), resourceGroupName, fleetName, top, skipToken, accept, context);
         return new PagedResponseBase<>(res.getRequest(), res.getStatusCode(), res.getHeaders(), res.getValue().value(),
             res.getValue().nextLink(), null);
     }
@@ -853,7 +883,9 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
     public PagedIterable<AutoUpgradeProfileInner> listByFleet(String resourceGroupName, String fleetName) {
-        return new PagedIterable<>(() -> listByFleetSinglePage(resourceGroupName, fleetName),
+        final Integer top = null;
+        final String skipToken = null;
+        return new PagedIterable<>(() -> listByFleetSinglePage(resourceGroupName, fleetName, top, skipToken),
             nextLink -> listByFleetNextSinglePage(nextLink));
     }
 
@@ -862,6 +894,8 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param fleetName The name of the Fleet resource.
+     * @param top The number of result items to return.
+     * @param skipToken The page-continuation token to use with a paged version of this API.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws ManagementException thrown if the request is rejected by server.
@@ -869,9 +903,9 @@ public final class AutoUpgradeProfilesClientImpl implements AutoUpgradeProfilesC
      * @return the response of a AutoUpgradeProfile list operation as paginated response with {@link PagedIterable}.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
-    public PagedIterable<AutoUpgradeProfileInner> listByFleet(String resourceGroupName, String fleetName,
-        Context context) {
-        return new PagedIterable<>(() -> listByFleetSinglePage(resourceGroupName, fleetName, context),
+    public PagedIterable<AutoUpgradeProfileInner> listByFleet(String resourceGroupName, String fleetName, Integer top,
+        String skipToken, Context context) {
+        return new PagedIterable<>(() -> listByFleetSinglePage(resourceGroupName, fleetName, top, skipToken, context),
             nextLink -> listByFleetNextSinglePage(nextLink, context));
     }